Replaced super_exception_notifier with exception_notification.
super_exception_notifier doesn't yet support rails3, so I changed to
the vanilla rails plugin at:
http://github.com/rails/exception_notification

There isn't a gem for it yet, so I built it and packed it with
bundler.

One disadvantage is that the vanilla plugin doesn't support giving the
notification additional data (like the current group). The advantage
is that it's rack-based and less intrusive, so easier for each shapado
user to replace later with Hoptoad, Exceptional or something else.