Localisation updates from https://translatewiki.net.
This is the last and full export from the translatewiki.net
platform. Development of Shapado appears to have halted over
a year ago, and it does not make sense to ask volunteer
translators to translate software that is no longer being
developed.

It's been fun while it lasted. Here it ends. On to new
projects!